+++ src/blocknode/block_node.hh
@@ -3,7 +3,11 @@
 #include "../nodes/node.hh"
 #include "local_io.hh"
 #include "../messages/IOoperation.hh"
+#include "../common/context_singleton.hh"
 #include <string>
+#include <map>
+#include <queue>
+
 
 namespace eclipse {
 
@@ -21,10 +25,12 @@ class BlockNode: public Node {
     bool block_delete_local(Block&, bool replicate = true);
 
     //! @brief Update the content of the block.
-    bool block_update_local(Block& block, uint32_t pos, uint32_t len, bool replicate = true);
+    //bool block_update_local(Block& block, uint32_t pos, uint32_t len, bool replicate = true);
+    bool block_update_local(Block& block, uint64_t pos, uint64_t len, bool replicate = true);
 
     //! @brief Read block from the local node.
     bool block_read_local(Block& block, uint64_t off = 0, uint64_t len = 0, bool ignore_params = true);
+
 
   protected:
     void replicate_message(messages::IOoperation*);
